New Stereo Parallax Mode. Activate with: mOmni.sphereRadius(radius_of_allosphere_in_opengl_units).
In the new mode, we use negative parallax inside the sphere, zero parallax on the sphere and positive parallax outside the sphere. The most clear difference is if you point a laser pointer, you’ll find it is actually at some distance instead of infinity, if you point to an object outside, you’ll feel the pointer in front of it, if you point an object inside, the pointer is behind it, and if the object is just at the sphere surface, the pointer is on it. In addition, this mode allows you to see objects that are on the sphere surface clearly without glasses.