Change drop callback to use std::string
This is mainly due to the strange way SDL implements a multi-file drop callback. Each file in the group calls the callback and, bizarrely, since memory needs to be freed for the provided SDL file path, we need to copy the string. The allo drop callback could be refactored as one file at a time like SDL, but then we won't know if a multi-file drop occurred.