CSVReader overhaul
This fixes a major limitation of the previous implementation which only supported fixed-size string field lengths. It is also gives the user more control in how fields are converted and stored in custom data structures (by using an iterator). This approach also defers data conversions to only when they are needed, potentially improving performance. Support was also added for double-quoted fields commonly used if the field contains a delimiter.