Nav: remove cached local frame
Caching the orientation frame adds a lot of complexity to the class as every function that modifies the internal quat needs to update the cache, o.w. the object gets in a bad state. This also leads to a proliferation of overloaded functions. With the cache removed, Nav can be treated more or less as a Pose. Efficiency is probably not a concern as step() needs to compute a new frame anyway. If efficiency becomes a concern, there could probably be added an overloaded step() that returns the computed frame.